+++ apps/admin/app/(main)/(super-admin)/super-admin/user-manage/page.tsx
@@ -28,7 +28,7 @@ export default async function Page({
       <main className="flex h-full flex-col grow gap-5 pl-16 pr-20 pt-20">
         <header className="flex items-center justify-between">
           <h1 className="text-[34px] font-bold">사용자 관리</h1>
-          <UserAddButton page={currentPage}/>
+          <UserAddButton page={currentPage} />
         </header>
         <Suspense fallback={<LoadingFallback />}>
           <UserTableSection />
