+++ apps/admin/app/providers.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { OverlayProvider } from "@toss/use-overlay";
 import { QueryClientProvider, ReactQueryDevtools } from "@uket/api";
 
 import { getQueryClient } from "@uket/api/get-query-client";
@@ -9,7 +10,7 @@ export default function Providers({ children }: { children: React.ReactNode }) {
 
   return (
     <QueryClientProvider client={queryClient}>
-      {children}
+      <OverlayProvider>{children}</OverlayProvider>
       <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>
   );
